requirements.txt
streamlit
sympy
pandas
numpy
scipy
supabase
fpdf
pillow
requests